requirements.txt
Django==2.1.5
django-filter==2.1.0
graphene-django==2.2.0
django-graphql-jwt==0.2.0